powerpc/perf/24x7:
Add a simple test case to test the transaction interface of perf
with 24x7 counters. This transaction functionality was discussed
in https://lkml.org/lkml/2015/9/3/658 and merged into v4.3. Eg:

	commit 4a00c16
	Author: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>
	Date:   Thu Sep 3 20:07:51 2015 -0700
	perf/core: Define PERF_PMU_TXN_READ interface

This test case uses the transaction interface to minimize the number
of hcalls the underlying kernel makes on PowerPC systems.  It creates
a single group of a set of 24x7 events. It then uses the read() system
call every 2 seconds to read them all at once and displays the counters.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>